+++ components/core/src/os/signals/mod.rs
@@ -17,25 +17,30 @@
 // our homespun implementation. Check for status of that here:
 // https://github.com/rust-lang/rfcs/issues/1368
 
-use crate::os::process;
-
-#[cfg(unix)]
-pub enum SignalEvent {
-    WaitForChild,
-    Passthrough(process::Signal),
-}
+use std::sync::atomic::{AtomicBool,
+                        Ordering};
 
 #[cfg(unix)]
 mod unix;
 
-#[cfg(windows)]
-mod windows;
-
 #[cfg(unix)]
-pub use self::unix::{check_for_shutdown,
-                     check_for_signal,
-                     init};
+pub use self::unix::{check_for_signal,
+                     init,
+                     SignalEvent};
 
+static SHUTDOWN: AtomicBool = AtomicBool::new(false);
+
+// NOTE: The Unix implementation of `init` also establishes a similar
+// handler for shutdown signals, but also does some other stuff, as
+// well. Seems best for now to keep all those implementation details
+// in the `unix` module.
 #[cfg(windows)]
-pub use self::windows::{check_for_shutdown,
-                        init};
+pub fn init() {
+    use ctrlc;
+    ctrlc::set_handler(move || {
+        SHUTDOWN.store(true, Ordering::SeqCst);
+    }).expect("Error setting Ctrl-C handler");
+}
+
+/// Returns `true` if we have received a signal to shut down.
+pub fn check_for_shutdown() -> bool { SHUTDOWN.compare_and_swap(true, false, Ordering::SeqCst) }
